Extract registry step summary helpers into shared module
