+++ workstation/notes/registry-report.json
@@ -1,6 +1,6 @@
 {
   "schemaVersion": 2,
-  "generatedAt": "2026-03-22T07:20:50.740Z",
+  "generatedAt": "2026-03-22T08:12:00.228Z",
   "counts": {
     "idExports": 974,
     "uniqueIdExports": 974,
