+++ workstation/notes/registry-report.md
@@ -1,6 +1,6 @@
 # Registry Data Quality Report
 
-Generated: 2026-03-22T07:20:50.740Z
+Generated: 2026-03-22T08:12:00.228Z
 
 ## Counts
 - ID exports: 974 (unique: 974)
